stydy: add test, change source directory, and refactor
