+++ README.md
@@ -11,3 +11,11 @@ deno task start
 ```
 
 Then you can access to http://localhost:8000/
+
+## 🧪 Test
+
+If you try the deno test, you could run below command.
+
+```sh
+deno test --allow-read --allow-env --allow-net
+```
