+++ src/routes/_500.tsx
