+++ src/routes/api/joke.ts
