+++ src/routes/api/random-uuid.ts
