+++ src/routes/partials/about/[id].tsx
