+++ src/routes/projects/[id].tsx
@@ -1,4 +1,5 @@
 import { Handlers, PageProps } from "$fresh/server.ts";
+import { Link } from "../../components/Link.tsx";
 
 const projects = [{ id: 1, name: "Project 1", stars: 10 }, {
   id: 2,
@@ -37,6 +38,9 @@ export default function ProjectPage(props: PageProps<Project>) {
     <div>
       <h1 class="text-4xl font-bold">{props.data.name}</h1>
       <p>{props.data.stars} stars</p>
+      <div class="mt-4">
+        <Link text="Back" href="/" color="secondary" />
+      </div>
     </div>
   );
 }
